refactor: make Docker image parser accept any digest format
Changed the parser to accept any string after @ as a digest, moving the
validation of digest format to the validation layer. This provides better
separation of concerns:

- Parser: Extracts structure from the input
- Validator: Checks if the structure is valid

This change allows the parser to successfully parse images with invalid
digest formats (like 'cimg/go:1.24@foo'), which can then be properly
validated and reported as errors in the validation phase.